Add library in Tyepscript
