+++ test/mocha.opts
@@ -1,0 +1,4 @@
+--require ts-node/register
+--require source-map-support/register
+--full-trace
+src/**/*.test.ts
