+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "declaration": true,
+    "outDir": "./lib",
+    "strict": true,
+    "lib": [ "es2015" ],
+    "types": ["node"]
+  },
+  "include": ["src"],
+  "exclude": ["node_modules", "**/tests/*"]
+}
