+++ tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["tslint:recommended", "tslint-config-prettier"]
+}
